Use term wiki wrapper.
